+++ lab_2/const.py
@@ -1,1 +1,3 @@
 PI = {1: 0.2148, 2: 0.3672, 3: 0.2305, 4: 0.1875}
+DIR = "lab_2"
+FILE = "source.json"
