+++ lab_2/main.py
@@ -1,12 +1,13 @@
 import os
 
 import nist_test
-import json_loader
+
+from const import DIR, FILE
 
 
 if __name__ == "__main__":
-    java_seq = json_loader.load(os.path.join("lab_2", "source.json"))["java"]
-    cpp_seq = json_loader.load(os.path.join("lab_2", "source.json"))["cpp"]
+    java_seq = nist_test.load(os.path.join(DIR,FILE))["java"]
+    cpp_seq = nist_test.load(os.path.join(DIR,FILE))["cpp"]
     print("Nist java test:")
     print(nist_test.frequency_bit_test(java_seq))
     print(nist_test.identical_consecutive_bits(java_seq))
